Bytter ut moduler med komponenter som er trukket ut

+++ app/src/main/kotlin/no/nav/aap/behandlingsflyt/behandling/avklaringsbehov/BehandlingTilstandValidator.kt
@@ -1,9 +1,9 @@
 package no.nav.aap.behandlingsflyt.behandling.avklaringsbehov
 
-import no.nav.aap.behandlingsflyt.dbconnect.DBConnection
 import no.nav.aap.behandlingsflyt.sakogbehandling.behandling.BehandlingRepositoryImpl
 import no.nav.aap.behandlingsflyt.sakogbehandling.behandling.flate.BehandlingReferanse
 import no.nav.aap.behandlingsflyt.sakogbehandling.behandling.flate.BehandlingReferanseService
+import no.nav.aap.komponenter.dbconnect.DBConnection
 import no.nav.aap.motor.FlytJobbRepository
 
 class BehandlingTilstandValidator(connection: DBConnection) {
@@ -14,7 +14,7 @@ class BehandlingTilstandValidator(connection: DBConnection) {
         val behandling = behandlingReferanseService.behandling(behandlingReferanse)
         ValiderBehandlingTilstand.validerTilstandBehandling(behandling, behandlingVersjon)
 
-        val jobberForBehandling = flytJobbRepository.hentJobberForBehandling(behandling.id)
+        val jobberForBehandling = flytJobbRepository.hentJobberForBehandling(behandling.id.toLong())
         if (jobberForBehandling.isNotEmpty()) {
             throw BehandlingUnderProsesseringException()
         }
